+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.4.0] - unreleased
 
 - Add support for interceptors
-- Add `by` decorator
+- Add `generate` decorator
 - Add `timestamp` generator
 - Fix getDependencies that fail to list all relative files